+++ crates/block-producer/src/test_utils/mod.rs
@@ -1,7 +1,4 @@
-use std::sync::Arc;
-
 use miden_objects::{
-    accounts::AccountId,
     crypto::rand::{FeltRng, RpoRandomCoin},
     transaction::TransactionId,
     Digest,