+++ crates/block-producer/src/test_utils/store.rs
@@ -1,19 +1,20 @@
 use std::{
     collections::{BTreeMap, BTreeSet},
     num::NonZeroU32,
+    sync::Arc,
 };
 
 use miden_node_proto::domain::{blocks::BlockInclusionProof, notes::NoteAuthenticationInfo};
 use miden_objects::{
+    accounts::AccountId,
     block::{Block, NoteBatch},
     crypto::merkle::{Mmr, SimpleSmt, Smt, ValuePath},
     notes::{NoteId, NoteInclusionProof, Nullifier},
     transaction::ProvenTransaction,
-    BlockHeader, ACCOUNT_TREE_DEPTH, EMPTY_WORD, ZERO,
+    BlockHeader, Digest, ACCOUNT_TREE_DEPTH, EMPTY_WORD, ZERO,
 };
 use tokio::sync::RwLock;
 
-use super::*;
 use crate::{
     batch_builder::TransactionBatch,
     block::{AccountWitness, BlockInputs},